Move CUDA SFM implementation and docs into sfm
