+++ gtsam/sfm/cuda/SfmLevenbergMarquardt.cu
@@ -28,12 +28,12 @@
 #include <gtsam/nonlinear/cuda/internal/PcgLmPolicy.h>
 #include <gtsam/nonlinear/internal/LevenbergMarquardtPolicy.h>
 #include <gtsam/slam/GeneralSFMFactor.h>
-#include <gtsam/slam/cuda/SfmLevenbergMarquardt.h>
-#include <gtsam/slam/cuda/internal/SfmDenseSchurSolver.h>
-#include <gtsam/slam/cuda/internal/SfmProjectionLinearization.h>
-#include <gtsam/slam/cuda/internal/SfmReducedCsrPlan.h>
-#include <gtsam/slam/cuda/internal/SfmSchurProblem.h>
-#include <gtsam/slam/cuda/internal/SfmValues.h>
+#include <gtsam/sfm/cuda/SfmLevenbergMarquardt.h>
+#include <gtsam/sfm/cuda/internal/SfmDenseSchurSolver.h>
+#include <gtsam/sfm/cuda/internal/SfmProjectionLinearization.h>
+#include <gtsam/sfm/cuda/internal/SfmReducedCsrPlan.h>
+#include <gtsam/sfm/cuda/internal/SfmSchurProblem.h>
+#include <gtsam/sfm/cuda/internal/SfmValues.h>
 
 #include <algorithm>
 #include <chrono>
