+++ gtsam/sfm/cuda/SfmTypes.h
