+++ gtsam/sfm/cuda/internal/SfmDenseSchurSolver.h
@@ -22,8 +22,8 @@
 #include <gtsam/dllexport.h>
 #include <gtsam/linear/cuda/LinearSolver.h>
 #include <gtsam/nonlinear/cuda/internal/DeviceValues.h>
-#include <gtsam/slam/cuda/internal/SfmProjectionBatch.h>
-#include <gtsam/slam/cuda/internal/SfmProjectionLinearization.h>
+#include <gtsam/sfm/cuda/internal/SfmProjectionBatch.h>
+#include <gtsam/sfm/cuda/internal/SfmProjectionLinearization.h>
 
 #include <cuda_runtime_api.h>
