+++ gtsam/sfm/cuda/internal/SfmReducedCsrPlan.h
