+++ gtsam/sfm/cuda/internal/SfmSchurOperator.cu
@@ -17,8 +17,8 @@
  */
 
 #include <gtsam/base/cuda/Errors.h>
-#include <gtsam/slam/cuda/internal/SfmSchurOperator.h>
-#include <gtsam/slam/cuda/internal/SfmSchurProblem.h>
+#include <gtsam/sfm/cuda/internal/SfmSchurOperator.h>
+#include <gtsam/sfm/cuda/internal/SfmSchurProblem.h>
 
 #include <cmath>
 #include <limits>
