+++ gtsam/slam/CMakeLists.txt
@@ -6,12 +6,5 @@ set (slam_excluded_headers #"")
 file(GLOB slam_headers "*.h")
 install(FILES ${slam_headers} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/slam)
 
-if(GTSAM_ENABLE_CUDA)
-  set(slam_cuda_headers
-      cuda/SfmLevenbergMarquardt.h
-      cuda/SfmTypes.h)
-  install(FILES ${slam_cuda_headers} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/slam/cuda)
-endif()
-
 # Build tests
 add_subdirectory(tests)
